create run command line service
