+++ src/main/kotlin/org/jetbrains/research/testspark/services/TestCoverageCollectorService.kt
@@ -34,32 +34,6 @@ class TestCoverageCollectorService(private val project: Project) {
     private val log = Logger.getInstance(this::class.java)
 
     /**
-     * Executes a command line process and returns the output as a string.
-     *
-     * @param cmd The command line arguments as an ArrayList of strings.
-     * @return The output of the command line process as a string.
-     */
-    fun runCommandLine(cmd: ArrayList<String>): String {
-        var errorMessage = ""
-
-        val process = ProcessBuilder()
-            .command("bash", "-c", cmd.joinToString(" "))
-            .redirectErrorStream(true)
-            .start()
-
-        val reader = BufferedReader(InputStreamReader(process.inputStream))
-        var line: String?
-
-        while (reader.readLine().also { line = it } != null) {
-            errorMessage += line
-        }
-
-        process.waitFor()
-
-        return errorMessage
-    }
-
-    /**
      * Generates the path for the command by concatenating the necessary paths.
      *
      * @param buildPath The path of the build file.
@@ -99,7 +73,7 @@ class TestCoverageCollectorService(private val project: Project) {
         // find the proper javac
         val javaCompile = File(javaHomeDirectory.path).walk().filter { it.name.equals("javac") && it.isFile }.first()
         // compile file
-        val errorMsg = runCommandLine(
+        val errorMsg = project.service<RunCommandLineService>().runCommandLine(
             arrayListOf(
                 javaCompile.absolutePath,
                 "-cp",
@@ -172,7 +146,7 @@ class TestCoverageCollectorService(private val project: Project) {
         name += "$className#$testCaseName"
 
         // run the test method with jacoco agent
-        val testExecutionError = runCommandLine(
+        val testExecutionError = project.service<RunCommandLineService>().runCommandLine(
             arrayListOf(
                 javaRunner.absolutePath,
                 "-javaagent:$jacocoAgentDir=destfile=$dataFileName.exec,append=false,includes=${project.service<Workspace>().classFQN}",
@@ -218,7 +192,7 @@ class TestCoverageCollectorService(private val project: Project) {
 
         log.info("Runs command: ${command.joinToString(" ")}")
 
-        runCommandLine(command as ArrayList<String>)
+        project.service<RunCommandLineService>().runCommandLine(command as ArrayList<String>)
 
         return testExecutionError
     }
